+++ test/reducers/game.spec.js
@@ -5,7 +5,7 @@ import * as cards from '../../src/common/store/cards';
 import {
   BLUE_CORE_HEX, ORANGE_CORE_HEX, STARTING_PLAYER_HEALTH, TYPE_ROBOT, TYPE_STRUCTURE
 } from '../../src/common/constants';
-import { getCost } from '../../src/common/util';
+import { getCost } from '../../src/common/util/game';
 import {
   getDefaultState, objectsOnBoardOfType, queryObjectAttribute, queryRobotAttributes, queryPlayerHealth,
   newTurn, drawCardToHand, playObject, playEvent, moveRobot, attack,
