+++ test/testHelpers.js
@@ -2,10 +2,8 @@ import game from '../src/common/reducers/game';
 import * as actions from '../src/common/actions/game';
 import { collection, attackBotCard } from '../src/common/store/cards';
 import defaultState from '../src/common/store/defaultGameState';
-import {
-  instantiateCard, allObjectsOnBoard, ownerOf, getAttribute,
-  drawCards
-} from '../src/common/util';
+import { instantiateCard } from '../src/common/util/common';
+import { allObjectsOnBoard, ownerOf, getAttribute, drawCards } from '../src/common/util/game';
 import { transportObject } from '../src/common/reducers/handlers/game/board';
 
 export function getDefaultState() {
